requirements.txt
pystache
docopt
pyyaml
stevedore
termcolor
